+++ src/components/NewsCard.vue
@@ -213,7 +213,7 @@ export default {
     categoryClick(e) {
       goToSearch(this.$router, {
         query: e.target.innerText.toLowerCase(),
-        type: "hub",
+        type: "general",
       });
     },
     errorHandler(id) {