+++ src/views/About/CompositionAndMembership.vue
@@ -79,7 +79,7 @@ export default {
       return renderToHtml(content);
     },
     search(name) {
-      goToSearch(this.$router, { query: name, type: "hub" });
+      goToSearch(this.$router, { query: name, type: "general" });
     },
   },
   apollo: {