+++ src/views/Grants/GrantsStaff.vue
@@ -59,7 +59,7 @@ export default {
       return renderToHtml(content);
     },
     search(name) {
-      goToSearch(this.$router, { query: name, type: "hub" });
+      goToSearch(this.$router, { query: name, type: "general" });
     },
   },
   apollo: {